backend/requirements.txt
Django>=4.2,<5.3
djangorestframework>=3.15
django-cors-headers>=4.4
dj-database-url>=2.2
gunicorn>=21.2
whitenoise>=6.7
psycopg2-binary>=2.9
